+++ Kronko/ColorPicker.cpp
@@ -3,7 +3,7 @@
 
 ColorPicker::ColorPicker(){}
 
-cv::Vec3i ColorPicker::getColorV(Mat & img, Point p, double radius)
+cv::Vec3i ColorPicker::getColorV(Mat & img, Point p, int radius)
 {
     std::cerr << "[ColorPicker::getColorV] Should'nt be called." << std::endl;
     return cv::Vec3i(0,0,0);
