+++ Kronko/ColorPoint.cpp
@@ -1,6 +1,6 @@
 #include "ColorPoint.h"
 
-Vec3i ColorPoint::getColorV(Mat &img, Point p, double radius)
+Vec3i ColorPoint::getColorV(Mat &img, Point p, int radius)
 {
 	Size sz = img.size();
 	if (p == Point(-1, -1)) p = Point(sz.width / 2, sz.height / 2);
